Update package versions and add a new one
This commit updates the version of the library in Directory.Build.props file from -preview2 to -preview3, ensuring the most recent features are included. It also comments out the OpenTelemetry.Contrib.Instrumentation package versions in the Directory.Packages.props file due to instability issues of the beta versions used earlier. A new package, Microsoft.DiaSymReader was added to the same file for better symbolism reading.

Signed-off-by: afroze9 <afrozeamjad@gmail.com>